Refactor within main services
